Mise en place du controle admin sur la requête model getAllModels

+++ ouiMouve/src/main/java/com/api/ouimouve/service/SiteService.java
@@ -110,12 +110,12 @@ public class SiteService {
         /**
          * Checks if the site is referenced by any carPooling departure or destination.
          */
-//        boolean siteUsedByCarPooling = carPoolingService.getCarPoolingByFilter(null, null,null,null,site.getName(),null, null).isEmpty() &&
-//                carPoolingService.getCarPoolingByFilter(null, null,null,
-//                        null,null,site.getName(), null, null).isEmpty();
-//        if (!siteUsedByCarPooling) {
-//            throw new InvalidRessourceException("Impossible de supprimer ce site : un ou plusieurs covoiturages le référence");
-//        }
+        boolean siteUsedByCarPooling = carPoolingService.getCarPoolingByFilter(null, null,null,null, site.getName(),null, null, null).isEmpty() &&
+                carPoolingService.getCarPoolingByFilter(null, null,null,
+                        null,null,site.getName(), null, null).isEmpty();
+        if (!siteUsedByCarPooling) {
+            throw new InvalidRessourceException("Impossible de supprimer ce site : un ou plusieurs covoiturages le référence");
+        }
         siteRepository.delete(site);
 
         return siteMapper.toSiteResponseDto(site);
